Split textUi into textUi and Formatter classes
